+++ lib/whois/record/parser/whois.nic.ms.rb
@@ -15,6 +15,10 @@ module Whois
     class Parser
 
       # Parser for the whois.nic.ms server.
+      #
+      # @see Whois::Record::Parser::Example
+      #   The Example parser for the list of all available methods.
+      #
       class WhoisNicMs < BaseCocca
       end