+++ spec/whois/record/parser/responses/whois.cmc.iq/status_registered_spec.rb
@@ -55,13 +55,13 @@ describe Whois::Record::Parser::WhoisCmcIq, "status_registered.expected" do
   describe "#updated_on" do
     it do
       subject.updated_on.should be_a(Time)
-      subject.updated_on.should == Time.parse("2012-05-24")
+      subject.updated_on.should == Time.parse("2012-10-17")
     end
   end
   describe "#expires_on" do
     it do
       subject.expires_on.should be_a(Time)
-      subject.expires_on.should == Time.parse("2012-10-03")
+      subject.expires_on.should == Time.parse("2013-10-03")
     end
   end
   describe "#registrar" do
@@ -69,6 +69,7 @@ describe Whois::Record::Parser::WhoisCmcIq, "status_registered.expected" do
       subject.registrar.should be_a(Whois::Record::Registrar)
       subject.registrar.id.should           == nil
       subject.registrar.name.should         == "CMC Registrar"
+      subject.registrar.organization.should == nil
       subject.registrar.url.should          == nil
     end
   end