+++ spec/whois/record/parser/responses/whois.nic.net.sb/status_registered_spec.rb
@@ -49,19 +49,19 @@ describe Whois::Record::Parser::WhoisNicNetSb, "status_registered.expected" do
   describe "#created_on" do
     it do
       subject.created_on.should be_a(Time)
-      subject.created_on.should == Time.parse("2010-03-27")
+      subject.created_on.should == Time.parse("2010-03-27 04:29:19 UTC")
     end
   end
   describe "#updated_on" do
     it do
       subject.updated_on.should be_a(Time)
-      subject.updated_on.should == Time.parse("2011-07-18")
+      subject.updated_on.should == Time.parse("2012-02-26 05:08:41 UTC")
     end
   end
   describe "#expires_on" do
     it do
       subject.expires_on.should be_a(Time)
-      subject.expires_on.should == Time.parse("2012-03-27")
+      subject.expires_on.should == Time.parse("2013-03-27 04:29:19 UTC")
     end
   end
   describe "#registrar" do
@@ -69,6 +69,7 @@ describe Whois::Record::Parser::WhoisNicNetSb, "status_registered.expected" do
       subject.registrar.should be_a(Whois::Record::Registrar)
       subject.registrar.id.should           == nil
       subject.registrar.name.should         == "Key-Systems"
+      subject.registrar.organization.should == nil
       subject.registrar.url.should          == "http://www.key-systems.net"
     end
   end