+++ CHANGELOG.md
@@ -11,6 +11,26 @@
 
 - CHANGED: Renamed Whois::Record::Scanners::Ast to Renamed Whois::Record::Scanners::Nodable
 
+- CHANGED: Updated whois.nic.af parser to the new response format.
+
+- CHANGED: Updated whois.nic.cx parser to the new response format.
+
+- CHANGED: Updated whois.nic.gs parser to the new response format.
+
+- CHANGED: Updated whois.nic.ht parser to the new response format.
+
+- CHANGED: Updated whois.nic.ki parser to the new response format.
+
+- CHANGED: Updated whois.nic.mu parser to the new response format.
+
+- CHANGED: Updated whois.nic.net.nf parser to the new response format.
+
+- CHANGED: Updated whois.nic.net.sb parser to the new response format.
+
+- CHANGED: Updated whois.nic.tl parser to the new response format.
+
+- CHANGED: Updated whois.registry.gy parser to the new response format.
+
 - FIXED: whois.domainregistry.ie should support status `Active - LOCKED`.
 
 - FIXED: whois.nic.uk fails to parse registrars without URL (GH-188).